jquery  / Editor and Display Templates
I made it so that anytime you use .EditorFor(m => m.SomeDate?) it
will display a textbox with only the date (no time), and will have jquery
datepicker built in.  Similar for DisplayFor, only this just renders a
string date with no time.

I also added a site.js file to Scripts that is now being referenced in the
App_Start\BundleConfig.cs.  In this .js I am applying the datepicker jquery
to any element that has the datepicker class.